host: Move libcontainer init code into flynn-host
